chore: create service layer for bulk module
