+++ src/server/inversify.config.ts
@@ -69,6 +69,8 @@ import { FileCheckController, UrlCheckController } from './modules/threat'
 
 import { QrCodeService } from './modules/qr/services'
 import { QrCodeController } from './modules/qr'
+
+import { BulkService } from './modules/bulk/services'
 import { BulkController } from './modules/bulk'
 
 function bindIfUnbound<T>(
@@ -136,6 +138,7 @@ export default () => {
   bindIfUnbound(DependencyIds.virusScanService, CloudmersiveScanService)
   bindIfUnbound(DependencyIds.fileCheckController, FileCheckController)
   bindIfUnbound(DependencyIds.bulkController, BulkController)
+  bindIfUnbound(DependencyIds.bulkService, BulkService)
 
   bindIfUnbound(DependencyIds.safeBrowsingMapper, SafeBrowsingMapper)
   bindIfUnbound(DependencyIds.safeBrowsingRepository, SafeBrowsingRepository)
