+++ src/server/modules/bulk/BulkController.ts
@@ -1,66 +1,18 @@
-import { bool } from 'aws-sdk/clients/signer'
 import { Request, Response } from 'express'
-import { injectable } from 'inversify'
-import paparse from 'papaparse'
-import { ogHostname } from '../../config'
-import { BULK_UPLOAD_HEADER, BULK_UPLOAD_MAX } from '../../constants'
-
-import * as validators from '../../../shared/util/validation'
-
+import { inject, injectable } from 'inversify'
 import jsonMessage from '../../util/json'
-
-type CSVSchema = {
-  rows: number
-  isValid: bool
-}
+import { DependencyIds } from '../../constants'
+import * as interfaces from './interfaces'
 
 @injectable()
 export class BulkController {
-  private parseCSV = (dataString: string): CSVSchema => {
-    const schema = {
-      rows: 0,
-      isValid: true,
-    } as CSVSchema
+  bulkService: interfaces.BulkService
 
-    paparse.parse(dataString, {
-      skipEmptyLines: false,
-      delimiter: ',',
-      step(step, parser) {
-        schema.rows += 1
-        const rowData = step.data as string[]
-        let validRow = true
-        if (schema.rows === 1) {
-          validRow = schema.rows === 1 && rowData[0] === BULK_UPLOAD_HEADER
-        } else {
-          const acceptableLinkCount = schema.rows <= BULK_UPLOAD_MAX + 1 // rows include header
-          const onlyOneColumn = rowData.length === 1
-          const isNotBlacklisted = !validators.isBlacklisted(rowData[0])
-          const isNotEmpty = rowData[0].length > 0
-          const isHttps = validators.isHttps(rowData[0])
-          const validCharacters = validators.isPrintableAscii(rowData[0])
-          const isNotCircularRedirect = !validators.isCircularRedirects(
-            rowData[0],
-            ogHostname,
-          )
-          const noParsingError = step.errors.length === 0
-          validRow =
-            isNotCircularRedirect &&
-            acceptableLinkCount &&
-            onlyOneColumn &&
-            isNotBlacklisted &&
-            isNotEmpty &&
-            isHttps &&
-            validCharacters &&
-            noParsingError
-        }
-
-        if (!validRow) {
-          schema.isValid = false
-          parser.abort()
-        }
-      },
-    })
-    return schema
+  public constructor(
+    @inject(DependencyIds.bulkService)
+    bulkService: interfaces.BulkService,
+  ) {
+    this.bulkService = bulkService
   }
 
   public csvValidation: (req: Request, res: Response) => Promise<void> = async (
@@ -79,9 +31,10 @@ export class BulkController {
       return
     }
 
-    const schema = this.parseCSV(file.data.toString())
+    const schema = this.bulkService.parseCsv(file.data.toString())
     if (schema.isValid) {
-      res.ok({ isValid: schema.isValid })
+      // res.ok({ isValid: schema.isValid })
+      // TODO: process CSV
     } else {
       res.badRequest({ isValid: schema.isValid })
     }
