+++ src/server/modules/bulk/interfaces/BulkService.ts
@@ -1,0 +1,8 @@
+export type CSVSchema = {
+  rows: number
+  isValid: boolean
+}
+
+export interface BulkService {
+  parseCsv(dataString: string): CSVSchema
+}
