+++ src/server/modules/bulk/services/index.ts
@@ -1,0 +1,1 @@
+export { BulkService, BulkService as default } from './BulkService'
